\name{implicit-package}
\alias{implicit-package}
\alias{implicit}
\docType{package}
\title{
\packageTitle{implicit}
}
\description{
\packageDescription{implicit}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{implicit}
\packageIndices{implicit}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{implicit}

Maintainer: \packageMaintainer{implicit}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
